Fix map extra spawning in Backrooms
The id used by the terrain was incorrect, but more importantly, map
extras were checked against the 'default' region, not the default
region.

Change the id for the map extra collection to match that used by the
terrains, and check the default region, not the region with the id
default for map extras.