Fix fishing spot check
In the conversion to use the generic flood-fill function, done in
d2eef9a72a501b2de53fbf4dbcdb2d5ab328a9aa, the predicate copied the
bounds check, but did not remove the negation. This means that any point
added to the flood-fill set must be outside of fishing bounds, not
inside them. Because the fishing bounds are drawn around the seed point,
this is never true for the seed point and the flood fill always
terminates immediately.

Additionally, add some bounds checks, because this function is called in
iuse::fish with bounds of 60, which allows it to go out of bounds.
And change that 60 to the constant for map size.

Co-authored-by: RenechCDDA <84619419+RenechCDDA@users.noreply.github.com>