Dino mod modifies region settings with copy-from
Simply extend the map extra lists instead of building a new region
settings for it.

Do the same for the cities! The cities overlay results in a bug where it
enables cities in the innawoods mod, but there are no cities to pick,
resulting in a crash. Remove the workaround for that.